**Add docker support** Fix various issues in the code
The changes in this commit fix several issues in the code.

The commit addresses issues with the .gitignore file, the cpuset.c file, the rc80211_minstrel.c file, the rate.c file, the cgroup-defs.h file, the ibss.c file, the cgroup_pids.c file, the cgroup.c file, and the percpu-refcount.h file.

These changes include fixing a missing newline in the .gitignore file, commenting out a line in the cpuset.c file, fixing a logical condition in the rc80211_minstrel.c file, fixing a logical condition in the rate.c file, adding a new function in the cgroup-defs.h file, fixing a logical condition in the ibss.c file, removing unnecessary code in the cgroup_pids.c file, fixing a logical condition in the cgroup.c file, and adding new functions in the percpu-refcount.h file.

These changes improve the functionality and maintainability of the code.